TINY-6891: Rename to listmodel according to naming convention
