+++ modules/tinymce/src/plugins/lists/main/ts/listmodel/Entry.ts
