+++ modules/tinymce/src/plugins/lists/main/ts/listmodel/NormalizeEntries.ts
