+++ modules/tinymce/src/plugins/lists/main/ts/listmodel/Util.ts
